Project 03_02 Successfully Forked.
After a mistake of my own that shall not be repeated, we were able to gather the old files and place them into the new repository for GitKracken to do commit pushes properly again.

As for what was done before, I have finished Project 03_02 with all of its code including the multitude of style rules in the stylesheets and the comments needed to explain my understanding on what new concepts were brought to use in this assignment.